server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai>=1.50.0
